# Feature/nodeless inquiry