… src/common/core/setcore.js
… test/common/core/setcore.spec.js