To speed up inquiries on core we implemented the functions in a manner when no sub-gmeNode objects are created during inquiry (the way of manipulating have not changed).
In the process we reviewed the setcore and make some changes:
- getOwnMemberPaths now available on the core API to get those members of a set, that has some information on the node's level (so not completely inherited)
- isFullyOverriddenMember now available on core API to check if a member have been added to the set of the node, though it was inherited as well
- a new optional flag 'overridden' is now used in the export format when describing the set members, to capture if a member was completely overridden on the node's inheritance level (so added to the set directly on the node)

To measure the speed increase we simply exported a relatively bigger model (189Kb in export size):
- with the current master branch, the speed was varying from 1605ms to 2038ms
- with the nodeless_inquiry addition, the speed was reduced to the range of 756ms to 1033ms
This is roughly a ***50%*** increase or we can say it is ***twice*** as fast as the original one. We have to mention that this increase may not effect (or not effect to this extent) all UI operation.