@lattmann 2015-10-17
What is the speed up by this change? Where are the measurements?