@kecso 2015-10-17
Thanks for the comment!
I've updated the PR with a quick measurement on the export (I've exported one version of an ADM project). Please let me know if you need any more details.